15.yamdb_final/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
requests
django==2.2.6
pytest
pytest-pythonpath==0.7.3
pytest-django
djangorestframework
django-filter==2.4.0
djangorestframework-simplejwt
asgiref==3.2.10
gunicorn==20.0.4
psycopg2-binary==2.9
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1 
django-environ